feat: 添加新版本设置弹窗壳子
